pull COFFEE-independent code out of coffee.py
