+++ tsfc/driver.py
@@ -12,7 +12,7 @@ from tsfc.quadrature import create_quadrature, QuadratureRule
 
 import coffee.base as coffee
 
-from tsfc import fem, gem as ein, impero as imp, scheduling as sch, optimise as opt
+from tsfc import fem, gem as ein, impero as imp, scheduling as sch, optimise as opt, impero_utils
 from tsfc.coffee import SCALAR_TYPE, generate as generate_coffee
 from tsfc.constants import default_parameters
 from tsfc.node import traversal
@@ -208,7 +208,11 @@ def compile_integral(integral, idata, fd, prefix, parameters):
         return None
     temporaries = make_temporaries(op for loop_indices, op in indexed_ops)
 
-    body = generate_coffee(indexed_ops, temporaries, shape_map, apply_ordering)
+    tree, indices, declare = impero_utils.process(indexed_ops,
+                                                  temporaries,
+                                                  shape_map,
+                                                  apply_ordering)
+    body = generate_coffee(temporaries, tree, indices, declare)
     body.open_scope = False
 
     funname = "%s_%s_integral_%s" % (prefix, integral_type, integral.subdomain_id())
